Structure: Restructure input
