+++ src/components/general/input/PescaAmountField/index.ts
@@ -1,0 +1,1 @@
+export * from './PescaInputField.component';
