+++ src/components/general/input/PescaButton/index.ts
@@ -1,0 +1,1 @@
+export * from './PescaButton.component';
