+++ src/components/general/input/PescaInputField/index.ts
@@ -1,0 +1,1 @@
+export * from './PescaInputField.component';
